+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 1.1.0 (2016-11-xx)
 
-* Added SendOperations methods accepting a String as recipient
+* Added SendOperations methods accepting a push notification type
 * Added Messeger#getUserProfile method with varargs fields argument
 * Upgraded RestFB dependency to latest 1.34.0
 * Updated default Facebook API version to 2.8